跳转无操作数基类的位置

+++ src/main/java/com/wjd/instructions/conversions/L2D.java
@@ -1,6 +1,6 @@
 package com.wjd.instructions.conversions;
 
-import com.wjd.instructions.constants.NoOperandsInstruction;
+import com.wjd.instructions.base.NoOperandsInstruction;
 import com.wjd.rtda.stack.Frame;
 
 /**
@@ -12,7 +12,7 @@ public class L2D extends NoOperandsInstruction {
     @Override
     public void execute(Frame frame) {
         long l = frame.getOpStack().popLong();
-        double d = (float) l;
+        double d = (double) l;
         frame.getOpStack().pushDouble(d);
     }
 }
